client/requirements.txt
streamlit==1.44
streamlit-aggrid==0.3.4.post3
streamlit-option-menu==0.3.6
streamlit-authenticator==0.3.2
bcrypt==4.0.1
pandas==2.2.3
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-dotenv==1.1.0
nest-asyncio==1.6.0
PyYAML==6.0.1
cryptography==42.0.5
requests==2.31.0

# Core LangChain and MCP
langchain==0.3.20
langchain-openai>=0.0.3
langchain-mcp-adapters==0.0.7
langgraph==0.3.30

# MCP stdio support for embedded servers
mcp>=1.0.2

# Enhanced AI Provider Support (optional packages)
# Note: Install these for enhanced mode functionality:
# pip install openai anthropic google-generativeai cohere mistralai

# Enhanced AI Provider Support
# Install these packages for full Enhanced Configuration functionality

# OpenAI (already included in base requirements via langchain-openai)
openai>=1.0.0

# Anthropic Claude
# anthropic>=0.25.0

# # Google Gemini
# google-generativeai>=0.5.0

# # Cohere
# cohere>=4.0.0

# # Mistral AI
# mistralai>=0.1.0

# Additional utilities for enhanced features
httpx>=0.25.0
aiohttp>=3.9.0
pydantic>=2.5.0